ISLE parser: remove scheme boolean literals
The scheme-style `#f` and `#t` syntax are shorthands for `0` and `1` in ISLE.
This is a bit misleading, since they are not actually booleans, but integers:
`#t` will match against any width integer with value `1`, but not against the
boolean value `true`. The correct way to match against a boolean value is to use
`$true` and `$false`.

These weren't being used anywhere other than in the tests, so I removed them
completely.

Copyright (c) 2024, Arm Limited.

Signed-off-by: Karl Meakin <karl.meakin@arm.com>