ISLE spec: update `<ident>` production in language-reference
The definition of the `<ident>` production in the language reference had
drifted out of sync with the actual implementation. Update the spec to
match the implementation

Also tighten up the implementation of `is_sym_first_char` so that
control characters (eg `\0`) are not considered symbol characters.

Copyright (c) 2024, Arm Limited.

Signed-off-by: Karl Meakin <karl.meakin@arm.com>